+++ tests/rust-e2e-fileops/src/main.rs
@@ -12,7 +12,7 @@ static FILE_CONTENTS: &str = "Hello, I am the file you're reading!";
 static FILE_PATH: &str = "/tmp/test_file.txt";
 
 fn create_test_file() {
-    println!(">> Creating /app");
+    println!(">> Creating test file");
 
     // let _ = DirBuilder::new().create("/app2").expect("Created /app");
 
@@ -109,6 +109,26 @@ fn test_pread() {
     };
 }
 
+fn test_pwrite() {
+    println!(">> test_pwrite");
+
+    let file = OpenOptions::new()
+        .write(true)
+        .create(true)
+        .open(FILE_PATH)
+        .expect("pwrite!");
+
+    let fd = file.as_raw_fd();
+
+    unsafe {
+        let data = CString::new("Hello, I am the file you're writing!").expect("valid C string");
+
+        let (buffer, length, _capacity) = data.into_bytes_with_nul().into_raw_parts();
+
+        assert_eq!(libc::pwrite(fd, buffer.cast(), length, 0), 37);
+    };
+}
+
 fn main() {
     create_test_file();
 
@@ -117,4 +137,5 @@ fn main() {
     test_open_read_contents();
     test_fgets();
     test_pread();
+    test_pwrite();
 }